// src/types.ts
export interface MatchFormData {
  name: string;
  age: string;
  food: string;
  pet: string;
  familyCount: string;
  country: string;
  education: string;
  religion: string;
  skinColor: string;
  height: string;
  weight: string;
  job: string;
  mbti: string;
  hobbies: string;
  relationshipGoals: string;
  smoking: string;
  languages: string;
  loveLanguage: string;
  livingPreference: string;
}